+++ src/main/java/com/inditex/challenge/exception/ErrorCode.java
@@ -1,7 +1,19 @@
 package com.inditex.challenge.exception;
 
-public class ErrorCode {
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
 
-    final public static String NEGATIVE_ID = "Id should be positive.";
-    final public static String BAD_DATE_FORMAT = "Date format should be yyyy-MM-dd-HH.mm.ss";
+@Getter
+public enum ErrorCode {
+
+    NEGATIVE_ID("Id should be positive", HttpStatus.BAD_REQUEST),
+    BAD_DATE_FORMAT("Date format should be yyyy-MM-dd-HH.mm.ss", HttpStatus.BAD_REQUEST);
+
+    private final String message;
+    private final HttpStatus status;
+
+    ErrorCode(String message, HttpStatus status) {
+        this.message = message;
+        this.status = status;
+    }
 }