+++ src/main/java/com/inditex/challenge/exception/GlobalExceptionHandler.java
@@ -1,5 +1,6 @@
 package com.inditex.challenge.exception;
 
+import com.inditex.challenge.exception.business.InvalidBrandException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
@@ -12,12 +13,17 @@ public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {
 
     @ExceptionHandler(Exception.class)
     public ResponseEntity<ResponseEntityBody> handlerException(Exception exception, WebRequest request) {
-        return parseBodyFromErrorCode(exception.getMessage(), 404);
+        return parseBodyFromErrorCode(exception.getMessage(), HttpStatus.BAD_REQUEST);
     }
 
-    private ResponseEntity<ResponseEntityBody> parseBodyFromErrorCode(String message, Integer status){
+    @ExceptionHandler(InvalidBrandException.class)
+    public ResponseEntity<ResponseEntityBody> handlerBrandException(InvalidBrandException exception, WebRequest request) {
+        return parseBodyFromErrorCode(exception.getMessage(), exception.getErrorCode().getStatus());
+    }
+
+    private ResponseEntity<ResponseEntityBody> parseBodyFromErrorCode(String message, HttpStatus status){
         ResponseEntityBody response = new ResponseEntityBody(message);
-        response.setErrorCode("400");
+        response.setErrorCode(String.valueOf(status.value()));
         return ResponseEntity.status(status).body(response);
     }
 }