+++ src/main/java/com/inditex/challenge/exception/business/InvalidBrandException.java
@@ -1,0 +1,7 @@
+package com.inditex.challenge.exception.business;
+
+import com.inditex.challenge.exception.ErrorCode;
+
+public class InvalidBrandException extends BusinessException {
+    public InvalidBrandException(ErrorCode errorCode) { super(errorCode); }
+}