+++ src/main/java/com/inditex/challenge/validator/ChainValidator.java
@@ -1,10 +1,20 @@
 package com.inditex.challenge.validator;
 
 import com.inditex.challenge.exception.ErrorCode;
+import com.inditex.challenge.exception.business.InvalidBrandException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 public class ChainValidator {
 
-    public static void validate(Integer id) throws Exception {
-        if (id < 0) throw new Exception("Exception while validating Product: " + ErrorCode.NEGATIVE_ID);
+    private static final Logger logger = LoggerFactory.getLogger(ChainValidator.class);
+
+    public static void validate(Integer id) throws InvalidBrandException {
+        logger.info("Chain validation begin.");
+
+        if (id < 0) {
+            logger.error("Exception while validating id: " + id);
+            throw new InvalidBrandException(ErrorCode.NEGATIVE_ID);
+        }
     }
 }