+++ src/main/java/com/inditex/challenge/validator/ProductValidator.java
@@ -1,10 +1,18 @@
 package com.inditex.challenge.validator;
 
 import com.inditex.challenge.exception.ErrorCode;
+import com.inditex.challenge.exception.business.InvalidProductException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 public class ProductValidator {
 
-    public static void validate(Integer id) throws Exception {
-        if (id < 0) throw new Exception("Exception while validating Product: " + ErrorCode.NEGATIVE_ID);
+    private static final Logger logger = LoggerFactory.getLogger(ProductValidator.class);
+
+    public static void validate(Integer id) throws InvalidProductException {
+        if (id < 0) {
+            logger.error("Invalid id: " + id);
+            throw new InvalidProductException(ErrorCode.NEGATIVE_ID);
+        }
     }
 }